// checkpoints/2026-04-24-before-table-map-selection/lib/cooperto/types.ts
export interface BookingCreateInput {
  date: string;
  time: string;
  pax: number;
  roomCode?: string;
  statusCode?: number;
  firstName: string;
  lastName: string;
  email?: string;
  phone?: string;
  note?: string;
  privacyAccepted: boolean;
  marketingAccepted: boolean;
}